Fix Error when creating empty RSS feed on PHP 8
When the feed is empty, ObjectIterator::size() throws a TypeError
exception on PHP 8.0+ (and a Warning on PHP 7.2+) due to calling count()
on $this->objectlist->objects, which is null in this case.

Initializing the ObjectList::$objects property to an empty array ensures
this never happens.

File was converted to UTF-8 and trailing ?> removed.

Signed-off-by: Damien Regad <dregad@mantisbt.org>